oops... sqlite support on the site is quite cool, but if you cannot generate the configuration freepbx sux0rz...
with this patch the retriev*.pl files now get the config from sqlite as well. I also fixed some wierd bugs on the files, and not if the connection 
to the sql cannot be done, the script will abort before writing any configuration (before it would leave you with a zero length configuration).

please test, works on my setup... but who knows..., I am only testing on sqlite :)