Automation of localization:
Now there is a scritp which scans the htdocs/admin directory for php files, and retrieves from it all the
messages, to create a POT file. Then, it will merge in the new strings into the existing translations. I have
merged the translations locally, and I see that many of them still contain old strings not used by freepbx, legacy of
good old amportal...

I will see if nothing has been broken, and then I commit the merged in strings.