Hardcode "/usr/bin/php" and "/bin/sh", as this is the way most Linux distros do this. I know this conflicts ticket:695, but I just don't see any good reason to make another process on every single command (/usr/bin/env and then php).
If a Linux system or *BSD people are having poblems, this can be deal on install time by patching the source, this will be the best for the running
systems (less process will be opend). This is only good for the final product. Sorry, but I am quite sure you already have a lot of patches for the
product anyway, one little more patch will not hurt.

The commands are:
{{{

 !#/bin/sh
 !#/usr/bin/php -q
}}}