Fix bugs in module activation code
There are two bugs here:  The first is an issue with hidden
modules and the special casing for the 'framework' module.
Make it more clear that anything with 'hidden' should be
skipped, and then there is the special case for 'framework'.

The second set of bugs is the lookup of already processed
settings.  The code attempts to use a specific array value,
rather than the name of the array in the call to "in_array()",
and that causes a run-time error.