Cleanup amportal script to run on FreeBSD
Remove unneeded trailing semicolons - while required in php,
they are unnecessary in shell.  Consistify if / then / else
processing throughout file.

Simplify logic for finding the freepbx.conf file - set it to
the expected value, and override that setting if needed.

Exit with an error code if freepbx_engine could not be found.