Simply logic when reported unsigned/tampered modules
Improve the quality of the generated email messages.
Only refer to the affected modules only as "modules" and not
"files" in some cases.  This allows for simplication of the
logic when generating the messages.  Only store pure text strings in
the notifications table, as those strings are used as
the basis of both email and html output.  Add some colons
to the emailed output, to more clearly show a list follows.