Set HOME to sane value when running gpg, not a temp dir
If HOME is set to something other than the directory that holds
the .gnupg directory to be used, gpg 2.1.8 will fail to operate
properly when the dirmngr daemon is spawned.  It's mildly
confusing that the gpg's command-line option for --homedir=
isn't really a HOME directory, it's the directory that holds
the actual gpg related files (typically $HOME/.gnupg).

On systems where the "asterisk" user's home directory in the
password file is not present (e.g.: "/nonexistant"), the FreePBX
code already has the notion of failing over to use ASTSPOOLDIR
for holding the .gnupg directory.  This change makes that code
work for real.