Rewrite dbhelper to use individual tables
There have been issues with kvstore using a single table to store all
its data. If that table is corrupted - for whatever reason - the entire
machine can be disabled.  This splits the kvstore into seperate tables,
so that even if one table is corrupt, the other tables will continue
to work.