Address FREEPBX-13343 Automated access to the setting command.
User reported when running fwconsole setting -r -q the command would hang.
Secondary issue the command defaulted to false.

Change: Moved from the deprecated dialog helper to the question helper
Change: Changed default to true.

Note the primary issue is not a bug. The -q opperator only suppresses outpuy
The question and dialog helper look for the -n flag to suppress interactivity.
So the user needs to pass -q to hide output and -n to supress questions.
Then the module with this patch assume "true" and reset the setting to the
default value.