FREEI-1952 Optimize the module functionIncLoaded method
Sometimes a modules include files are loaded outside of the Module
class. The Module class makes a best effort to identify when this
happens and prevents an include of the functions again, as it should.
The problem is, when an include has been found to be loaded, it never
updates its cache to tell itself to skip all of the checks next time.

This commit now flags a file as included once it has been found to
be included to provide some optimization and as much as a 21% speedup
in some scripts. Also, we have changed the way we are looking up if
a module has been loaded from using an in_array() check to a hash
table check using isset(). Isset() was far faster to use in my tests.