Add reset hooks for unit tests
Typically when unit testing modules inside of FreePBX, we find ourselves
mocking core dependencies. This is problemamtic, because once we mock a
module, we can no longer use the actual module in tests. To fix this, we
have created a reset hooks that can be called on the classes that handle
dependencies. This allows us to reset the FreePBX god object so that it
will reload dependencies using the standard process.