Update PHP command to full path in amportal script
The php command when used to check for the version installed used relative path notation and thus would mean the script would use the executors path variable to find the php binary. By changing it to the full path; it will now ensure that only the right php binary is executed.